Update commit
Recode some phrases and added 'unusual' JavaDocs to it
